+++ src/components/common/MainHeader.vue
@@ -49,8 +49,10 @@
                   </div>
                   <div class="btn-action-wrap">
                     <template v-if="item.type === 'invite'">
-                      <button class="btn-accept" @click="handleAcceptInvitation(item.id)">수락</button>
-                      <button class="btn-reject" @click="handleRejectInvitation(item.id)">거절</button>
+                      <button class="btn-accept" @click="handleAcceptInvitation(item.invitationId)">
+                        {{ item.invitationId }}
+                        수락</button>
+                      <button class="btn-reject" @click="handleRejectInvitation(item.invitationId)">거절</button>
                     </template>
                     <template v-else>
                       <button class="btn-delete" @click="handleDeleteNotification(item.id)">삭제</button>
@@ -68,7 +70,7 @@
               <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-chat-dots-fill" viewBox="0 0 16 16">
                 <path d="M16 8c0 3.866-3.582 7-8 7a9 9 0 0 1-2.347-.306c-.584.296-1.925.864-4.181 1.234-.2.032-.352-.176-.273-.362.354-.836.674-1.95.77-2.966C.744 11.37 0 9.76 0 8c0-3.866 3.582-7 8-7s8 3.134 8 7M5 8a1 1 0 1 0-2 0 1 1 0 0 0 2 0m4 0a1 1 0 1 0-2 0 1 1 0 0 0 2 0m3 1a1 1 0 1 0 0-2 1 1 0 0 0 0 2"/>
               </svg>
-              <em>99</em>
+              <em></em>
             </button>
           </div>
           <!-- 마이페이지 아이콘: 페이지 이동 -->
