+++ src/pages/chat/ChatPage.vue
@@ -1,474 +1,18 @@
-
 <template>
-  <section class="chat-room">
-    <div class="chat-room-inner">
-      <!-- 좌측: 채팅방 목록(컴포넌트) -->
-      <ChatSidebar
-          :chatRooms="chatRooms"
-          :activeRoomId="activeRoomId"
-          @select="selectRoom"
-      />
-
-      <!-- 우측: 채팅 메인 OR 초대 화면 -->
-      <!-- 초대 모드일 때 (SSOT 원칙 적용) -->
-      <InviteView
-          v-if="showInviteView"
-          :users="availableUsers"
-          :invitedUserIds="invitedUserIds"
-          :searchQuery="searchQuery"
-          @back="closeInviteView"
-          @invite="handleUserInvite"
-          @search="handleInviteSearch"
-      />
-
-      <!-- 일반 채팅 모드일 때 -->
-      <div class="chat-main" v-else>
-        <!-- 상단 헤더 -->
-        <div class="chat-header">
-          <div class="chat-room-name">
-            <img class="avatar" :src="currentRoom.avatar" alt="아바타" />
-            <span class="nickname">{{ currentRoom.nickname }}</span>
-          </div>
-          <div class="chat-header-menu" ref="menuContainer">
-            <button class="btn-more" @click="toggleMenu">
-              <i class="bi bi-three-dots"></i>
-            </button>
-            <div class="menu-dropdown" v-if="menuOpen" ref="menuDropdown">
-              <button class="btn-invite" @click="openInvite">초대하기</button>
-              <div class="divider"></div>
-              <button class="btn-exit" @click="leaveRoom">나가기</button>
-            </div>
-          </div>
-        </div>
-
-        <!-- 채팅 메시지 목록 -->
-        <div class="chat-content" ref="chatContentRef">
-          <div
-              class="chat-message"
-              v-for="msg in currentRoom.messages"
-              :key="msg.id"
-          >
-            <img class="avatar" :src="msg.avatar" alt="아바타" />
-            <div class="message-info">
-              <div class="message-top">
-                <span class="nickname">{{ msg.nickname }}</span>
-                <span class="time">{{ msg.time }}</span>
-              </div>
-              <p class="message-text">{{ msg.text }}</p>
-            </div>
-          </div>
-        </div>
-
-        <!-- 채팅 입력창 -->
-        <div class="chat-input-wrap">
-          <input
-              type="text"
-              class="chat-input"
-              placeholder="채팅을 입력하세요."
-              v-model="input"
-              @keyup.enter="sendMessage"
-          />
-          <button class="btn-send" @click="sendMessage">
-            <i class="bi bi-vector-pen"></i>
-          </button>
-        </div>
-      </div>
-    </div>
-  </section>
+  <div class="page-container">
+    <ChatRoom />
+  </div>
 </template>
 
 <script setup>
-import ChatSidebar from '@/components/chat/ChatSidebar.vue'
-import InviteView from '@/components/chat/InviteView.vue'
-import Avatar from '@/assets/img/0.png'
-import { mockUsers } from '@/data/mockUsers.js'
-import { ref, computed, onMounted, onUnmounted, nextTick, watch } from 'vue'
-
-// ===== 채팅방 관련 데이터 (기존) =====
-const chatRooms = ref([
-  {
-    id: 1,
-    avatar: Avatar,
-    nickname: '고라니',
-    time: '방금',
-    preview: '기말고사 망했는데 어떡하죠',
-    unread: 9,
-    messages: [
-      { id: 1, avatar: Avatar, nickname: '고라니', time: '오전 9:53', text: 'ㄷㄷㄷㄷㄷㄷㄷ' },
-    ]
-  },
-  {
-    id: 2,
-    avatar: Avatar,
-    nickname: 'ddd',
-    time: '오전 9:00',
-    preview: '오늘 점심 뭐먹지',
-    unread: 0,
-    messages: [
-      { id: 1, avatar: Avatar, nickname: 'ddd', time: '오전 8:15', text: 'ㅋㅋㅋㅋ' }
-    ]
-  }
-])
-
-const activeRoomId = ref(chatRooms.value[0].id)
-const input = ref('')
-const menuOpen = ref(false)
-const menuContainer = ref(null)
-const menuDropdown = ref(null)
-const chatContentRef = ref(null) // 채팅 컨텐츠 영역 ref 추가
-
-// ===== 초대 관련 데이터 (SSOT 원칙 적용) =====
-const showInviteView = ref(false)
-const searchQuery = ref('')
-const invitedUserIds = ref([]) // 초대된 사용자 ID 관리
-
-// 사용자 목록 (테스트용 더미 데이터)
-const availableUsers = ref(mockUsers)
-
-// ===== Computed =====
-const currentRoom = computed(() =>
-    chatRooms.value.find(room => room.id === activeRoomId.value) || chatRooms.value[0]
-)
-
-// ===== 기존 채팅 관련 함수들 =====
-
-// 메시지 전송
-function sendMessage() {
-  if (input.value.trim() !== '') {
-    currentRoom.value.messages.push({
-      id: Date.now(),
-      avatar: Avatar,
-      nickname: currentRoom.value.nickname,
-      time: new Date().toLocaleTimeString(),
-      text: input.value
-    })
-    input.value = ''
-    
-    // 메시지 추가 후 스크롤을 맨 아래로
-    scrollToBottom()
-  }
-}
-
-// 메뉴 외부 클릭 감지
-function handleClickOutside(event) {
-  if (menuOpen.value && menuContainer.value && !menuContainer.value.contains(event.target)) {
-    menuOpen.value = false
-  }
-}
-
-// ESC 키 처리
-function handleEscapeKey(event) {
-  if (event.key === 'Escape') {
-    if (showInviteView.value) {
-      closeInviteView()
-    } else if (menuOpen.value) {
-      menuOpen.value = false
-    }
-  }
-}
-
-// 채팅방 선택
-function selectRoom(id) {
-  activeRoomId.value = id
-  showInviteView.value = false // 방 선택시 초대 화면 닫기
-  searchQuery.value = '' // 검색어 초기화
-  
-  // 채팅방 변경 후 스크롤을 맨 아래로
-  nextTick(() => {
-    scrollToBottom()
-  })
-}
-
-// 스크롤을 맨 아래로 이동
-function scrollToBottom() {
-  nextTick(() => {
-    if (chatContentRef.value) {
-      chatContentRef.value.scrollTop = chatContentRef.value.scrollHeight
-    }
-  })
-}
-
-// 메뉴 토글
-function toggleMenu() {
-  menuOpen.value = !menuOpen.value
-}
-
-// 채팅방 나가기
-function leaveRoom() {
-  menuOpen.value = false
-  // 채팅방 나가기 로직
-}
-
-// ===== 초대 관련 함수들 =====
-
-// 초대 화면 열기
-function openInvite() {
-  menuOpen.value = false
-  showInviteView.value = true
-}
-
-// 초대 화면 닫기
-function closeInviteView() {
-  showInviteView.value = false
-  searchQuery.value = '' // 검색어 초기화
-}
-
-// 사용자 검색
-function handleInviteSearch(query) {
-  searchQuery.value = query
-  // TODO: 추후 실제 검색 API 연동시 구현
-}
-
-// 사용자 초대 처리
-function handleUserInvite(user) {
-  // 임시 처리: 초대된 사용자 목록에 추가
-  invitedUserIds.value.push(user.id)
-  console.log(`${user.nickname}님을 초대했습니다.`)
-
-  // TODO: 추후 실제 API 연동시 구현
-}
-
-// ===== 라이프사이클 훅 =====
-
-// 컴포넌트 마운트시 이벤트 리스너 등록
-onMounted(() => {
-  document.addEventListener('click', handleClickOutside)
-  document.addEventListener('keydown', handleEscapeKey)
-  
-  // 초기 스크롤을 맨 아래로
-  scrollToBottom()
-})
-
-// 현재 채팅방의 메시지가 변경될 때 스크롤 자동 이동
-watch(
-  () => currentRoom.value.messages,
-  () => {
-    scrollToBottom()
-  },
-  { deep: true }
-)
-
-// 컴포넌트 언마운트시 이벤트 리스너 제거 (메모리 누수 방지)
-onUnmounted(() => {
-  document.removeEventListener('click', handleClickOutside)
-  document.removeEventListener('keydown', handleEscapeKey)
-})
+import ChatRoom from '@/components/chat/ChatRoom.vue'; // ChatRoom 컴포넌트의 경로
 </script>
 
-<style lang="scss">
-// 메인 채팅방 컴포넌트 스타일
-// src > assets > scss > style.scss 임포트 필수
-@import "@/assets/scss/style.scss";
-
-.chat-room {
-  .chat-room-inner {
-    display: flex;
-    gap: 9px; // 좌측 사이드바와 우측 메인 사이의 간격
-
-    .chat-main {
-      width: 794px;
-      height: 778px;
-      border: 1px solid $dim-gray;
-      border-radius: 15px 15px 0px 0px;
-      background: $white;
-      display: flex;
-      flex-direction: column;
-      position: relative;
-
-      // 상단 헤더
-      .chat-header {
-        height: 72px;
-        padding: 16px;
-        border-bottom: 1px solid $dim-gray;
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        position: relative;
-
-        .chat-room-name {
-          display: flex;
-          align-items: center;
-          gap: 14px;
-
-          .avatar {
-            width: 40px;
-            height: 40px;
-            border-radius: 50%;
-            object-fit: cover;
-          }
-
-          .nickname {
-            font-family: $secondary-kr;
-            font-weight: 500;
-            font-size: 18px;
-            line-height: 1.25;
-            color: $dark-black;
-          }
-        }
-
-        .chat-header-menu {
-          position: relative;
-          .btn-more {
-            width: 26px;
-            height: 26px;
-            background: none;
-            border: none;
-            cursor: pointer;
-            color: $black;
-            font-size: 18px;
-            display: flex;
-            align-items: center;
-            justify-content: center;
-          }
-
-          .menu-dropdown {
-            position: absolute;
-            top: 30px;
-            right: -18px;
-            width: 64px;
-            height: 64px;
-            background: $white;
-            border: 1px solid $dim-gray;
-            border-radius: 5px;
-            box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
-            z-index: 10;
-
-            .btn-invite,
-            .btn-exit {
-              width: 100%;
-              padding: 7px;
-              background: none;
-              border: none;
-              font-family: $secondary-kr;
-              font-weight: 500;
-              font-size: 12px;
-              line-height: 1.25;
-              color: $black;
-              cursor: pointer;
-              text-align: center;
-
-              &:hover {
-                background-color: rgba($black, 0.05);
-              }
-            }
-
-            .divider {
-              height: 1px;
-              background-color: $dim-gray;
-              margin: 0 11px;
-            }
-          }
-        }
-      }
-
-      // 채팅 메시지 영역
-      .chat-content {
-        flex: 1;
-        padding: 16px;
-        overflow-y: auto;
-        display: flex;
-        flex-direction: column;
-        gap: 25px; // 메시지 간격
-
-        .chat-message {
-          display: flex;
-          gap: 21px;
-          align-items: flex-start;
-
-          .avatar {
-            width: 50px;
-            height: 50px;
-            border-radius: 50%;
-            object-fit: cover;
-            flex-shrink: 0;
-          }
-
-          .message-info {
-            flex: 1;
-
-            .message-top {
-              display: flex;
-              align-items: center;
-              gap: 13px;
-              margin-bottom: 7px;
-
-              .nickname {
-                font-family: $secondary-kr;
-                font-weight: 500;
-                font-size: 18px;
-                line-height: 1.25;
-                color: $dark-black;
-              }
-
-              .time {
-                font-family: $primary-kr;
-                font-weight: 350;
-                font-size: 12px;
-                line-height: 1.2;
-                color: $silver-black;
-              }
-            }
-
-            .message-text {
-              font-family: $primary-kr;
-              font-weight: 400;
-              font-size: 14px;
-              line-height: 1.2;
-              color: $dark-black;
-              margin: 0;
-              word-wrap: break-word;
-            }
-          }
-        }
-      }
-
-      // 채팅 입력창
-      .chat-input-wrap {
-        height: 60px;
-        border-top: 1px solid $dim-gray;
-        background-color: $light-gray;
-        display: flex;
-        align-items: center;
-        position: relative;
-
-        .chat-input {
-          flex: 1;
-          height: 100%;
-          border: none;
-          background: transparent;
-          padding: 20px;
-          font-family: $primary-kr;
-          font-weight: 350;
-          font-size: 14px;
-          line-height: 1.2;
-          color: $dark-black;
-          outline: none;
-
-          &::placeholder {
-            color: $silver-black;
-          }
-        }
-
-        .btn-send {
-          width: 60px;
-          height: 60px;
-          background-color: $main-color;
-          border: none;
-          cursor: pointer;
-          display: flex;
-          align-items: center;
-          justify-content: center;
-
-          &:hover {
-            background-color: darken($main-color, 10%);
-          }
-          .bi {
-            color: $white;
-            font-size: 30px;
-          }
-        }
-      }
-    }
-  }
+<style scoped>
+.page-container {
+  /* 페이지 전체를 감싸는 스타일에 맞게 조정 */
+  display: flex;
+  justify-content: center;
+  padding: 20px;
 }
 </style>
